+++ include/nfx/Graphics.h
@@ -36,6 +36,7 @@
 #include "nfx/graphics/gl/pipeline/passes/ImagePlanePass.h"
 #include "nfx/graphics/gl/pipeline/passes/OutlinePass.h"
 #include "nfx/graphics/gl/pipeline/passes/PointShadowPass.h"
+#include "nfx/graphics/gl/pipeline/passes/Polygon2DPass.h"
 #include "nfx/graphics/gl/pipeline/passes/PostProcessPass.h"
 #include "nfx/graphics/gl/pipeline/passes/PresentPass.h"
 #include "nfx/graphics/gl/pipeline/passes/RenderPass.h"